requirements.txt
setuptools
wheel
pillow
image
gitpython
ttkthemes